refactor: move videostore stuff to ./videostore
